central_server/requirements.txt
ImageHash==4.3.2
numpy==2.3.1
opencv_python==4.8.1.78
Pillow==11.2.1
protobuf==6.31.1
pycryptodomex==3.20.0
python-dotenv==1.1.0
qrcode==7.4.2
Requests==2.32.4
websockets==15.0.1
google==3.0.0